c file

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
 int main(void)
 {
 	int count =  1;
-	unsigned long long sum = 1, prev = 0,tmp;
+	unsigned long sum = 1, prev = 0,tmp;
 
 	while (count <= 50)
 	{
